test: split ProcFs contract suite
